use static handler rather than tripperware
