+++ pkg/querier/tripperware/roundtrip_test.go
@@ -3,17 +3,13 @@ package tripperware
 import (
 	"bytes"
 	"context"
-	"encoding/json"
 	"io"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
-	"runtime"
 	"testing"
 
 	"github.com/go-kit/log"
-	"github.com/prometheus/common/version"
-	v1 "github.com/prometheus/prometheus/web/api/v1"
 	"github.com/stretchr/testify/require"
 	"github.com/weaveworks/common/user"
 )
@@ -85,19 +81,6 @@ func TestRoundTrip(t *testing.T) {
 		next: http.DefaultTransport,
 	}
 
-	version.Version = "v1.14.0"
-	resp := &buildInfoResponse{
-		Status: "success",
-		Data: &v1.PrometheusVersion{
-			Version:   version.Version,
-			GoVersion: runtime.Version(),
-		},
-	}
-	expectedResp, err := json.Marshal(resp)
-	require.NoError(t, err)
-
-	buildInfoRoundTripper := NewBuildInfoRoundTripper()
-
 	middlewares := []Middleware{
 		MiddlewareFunc(func(next Handler) Handler {
 			return mockMiddleware{}
@@ -110,14 +93,12 @@ func TestRoundTrip(t *testing.T) {
 		middlewares,
 		mockCodec{},
 		mockCodec{},
-		buildInfoRoundTripper,
 	)
 
 	for _, tc := range []struct {
 		path, expectedBody string
 	}{
 		{"/foo", "bar"},
-		{"/api/v1/status/buildinfo", string(expectedResp)},
 		{queryExemplar, "bar"},
 		{queryRange, responseBody},
 		{query, responseBody},
